requirements.txt
colorama==0.4.4
python-dotenv==0.21.0
requests==2.25.1
